individual 2/27/2015

+++ Escape/src/byui/cit260/escape/control/InventoryControl.java
@@ -10,43 +10,42 @@ package byui.cit260.escape.control;
  * @author Kale
  */
 public class InventoryControl {
-    
-    public static double calcRaftCompletion(double logsneeded, double logsininventory, double ropeneeded, double ropeininventory, double storageneeded, double storageininventory){
-        
-        if (logsneeded < 10 || logsneeded > 100){
-		 
+
+    public static double calcRaftCompletion(double logsneeded, double logsininventory, double ropeneeded, double ropeininventory, double storageneeded, double storageininventory) {
+
+        if (logsneeded < 10 || logsneeded > 100) {
+
             return -1;
         }
-        if (logsininventory  <0 || logsininventory > 100){
-		
+        if (logsininventory < 0 || logsininventory > 100) {
+
             return -1;
         }
-        if (ropeneeded  <60 || ropeneeded > 600){
-		
+        if (ropeneeded < 60 || ropeneeded > 600) {
+
             return -1;
         }
-        if (ropeininventory  <0 || ropeininventory > 600){
-		 
+        if (ropeininventory < 0 || ropeininventory > 600) {
+
             return -1;
         }
-        if (storageneeded  <2 || storageneeded > 29){
-		
+        if (storageneeded < 2 || storageneeded > 29) {
+
             return -1;
         }
-        if (storageininventory  <0 || storageininventory > 29){
-		
+        if (storageininventory < 0 || storageininventory > 29) {
+
             return -1;
         }
-        
-         
+
         double needforcom = logsneeded + ropeneeded + storageneeded;
 
         double raftcom = needforcom - (needforcom - (logsininventory + ropeininventory + storageininventory));
         double persentcom = (raftcom / needforcom) * 100;
-	return persentcom;
+        return persentcom;
     }
-    
-      public int calRaftSize(int people, int crates) {
+
+    public double calRaftSize(double people, double crates) {
 
         if (people < 1 || people > 9) {
             return -1;
@@ -55,8 +54,8 @@ public class InventoryControl {
             return -1;
         }
 
-        int pspace = (6 * 10) * people;
-        int raftsize = pspace + (crates * 2);
+        double pspace = (6 * 10) * people;
+        double raftsize = pspace + (crates * 2);
         return raftsize;
     }
